+++ packages/spacecat-shared-data-access/src/models/base/base.collection.js
@@ -524,6 +524,12 @@ class BaseCollection {
       const updates = items.map((item) => item.record);
       const response = await this.entity.put(updates).go();
 
+      const now = new Date().toISOString();
+      items.forEach((item) => {
+        const { record } = item;
+        record.updatedAt = now;
+      });
+
       if (isNonEmptyArray(response.unprocessed)) {
         this.log.error(`Failed to process all items in batch write for [${this.entityName}]: ${JSON.stringify(response.unprocessed)}`);
       }