+++ packages/spacecat-shared-data-access/src/models/index.d.ts
@@ -28,6 +28,7 @@ export type * from './scrape-job';
 export type * from './scrape-url';
 export type * from './site';
 export type * from './site-candidate';
+export type * from './site-top-form';
 export type * from './site-top-page';
 export type * from './suggestion';
 export type * from './report';