+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-google-client-v1.4.39](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.38...@adobe/spacecat-shared-google-client-v1.4.39) (2025-07-27)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#878](https://github.com/adobe/spacecat-shared/issues/878)) ([b049828](https://github.com/adobe/spacecat-shared/commit/b04982839c0ff5e4de4ab0e37508c5eb5272a679))
+
 # [@adobe/spacecat-shared-google-client-v1.4.38](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.4.37...@adobe/spacecat-shared-google-client-v1.4.38) (2025-07-19)